Block deletion when active purschases exist
This matches the behavior of v1. I've also split all there separate flows into separate compoenents for clarity
