+++ client/dashboard/me/profile-deletion/test/account-deletion-modal.test.tsx
@@ -4,7 +4,8 @@
 import { screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { render } from '../../../test-utils';
-import AccountDeletionConfirmModal from '../confirmation-modal';
+import AccountDeletionConfirmModal from '../account-deletion-modal';
+import type { Purchase } from '@automattic/api-core';
 
 const defaultProps = {
 	onClose: jest.fn(),
@@ -92,5 +93,22 @@ describe( 'AccountDeletionConfirmModal', () => {
 		fireEvent.click( screen.getByText( 'Delete account' ) );
 
 		expect( onConfirm ).toHaveBeenCalled();
+	} );
+
+	it( 'blocks deletion when user has active purchases', async () => {
+		const { queryClient } = render( <AccountDeletionConfirmModal { ...defaultProps } /> );
+		// Seed React Query cache with one cancelable purchase
+		const { userPurchasesQuery } = require( '@automattic/api-queries' );
+		queryClient.setQueryData( userPurchasesQuery().queryKey, [
+			{
+				ID: 1,
+				expiry_status: 'active',
+				product_slug: 'pro_plan',
+				is_cancelable: true,
+			} as Partial< Purchase >,
+		] );
+		// React Query will notify subscribers; wait for the UI to update
+		await screen.findByText( 'You still have active purchases on your account.' );
+		expect( screen.getByText( 'Manage purchases' ) ).toBeInTheDocument();
 	} );
 } );
